refactor(tracing): split tracing.rs into submodules
Splits the ~5,594-line crate-root `tracing.rs` into focused submodules,
behavior-preserving:

  - tracing/register.rs       NetEventRegister trait + DynamicRegister /
                              EventRegister / NetLogMessage / event-log machinery
  - tracing/event_kind.rs     EventKind + all event types and their impls
  - tracing/metrics_client.rs metrics-server client + OTEventRegister
  - tracing/tracer.rs         tracing setup / log-file management (the former
                              inline `pub mod tracer`)

The root keeps the existing submodule decls (aof / event_aggregator /
telemetry / state_verifier), `state_hash_*`, and shared glue. Cross-module
calls are `pub(super)`; `super::` paths in moved code that resolved to the
crate root become `super::super::`.
